+++ Source/CesiumRuntime/Private/LoadGltfResult.h
@@ -155,7 +155,6 @@ struct LoadedPrimitiveResult {
    * for computing the UV at a hit location on a primitive.
    */
   CesiumGltf::IndexAccessorType IndexAccessor;
-
 #pragma endregion
 };